feat : 회원 & 비회원 페이지 방문 로그

+++ src/main/java/kr/or/ddit/account/lgn/service/LoginLogVO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class LoginLogVO {
 	
 	private int memId;
-	private int llId;
+	private Long llId;
 	private Date llCreatedAt;
 	private String llStatus;
